chore: Add Terraform configuration for DigitalOcean domain and records with reserved IP
This commit adds Terraform configuration for creating a DigitalOcean domain and records. It also includes the addition of a reserved IP for the domain. The domain name is set to "universiteams.com" and the records are created for "www.api", "www", "api", and "@". The IP address used for the records is the reserved IP associated with the DigitalOcean droplet.